+++ plugins/migratecmd/automigrate.go
@@ -136,3 +136,19 @@ func (p *plugin) getCachedCollections() (map[string]*models.Collection, error) {
 
 	return result, nil
 }
+
+func (p *plugin) hasCustomMigrations() bool {
+	files, err := os.ReadDir(p.options.Dir)
+	if err != nil {
+		return false
+	}
+
+	for _, f := range files {
+		if f.IsDir() {
+			continue
+		}
+		return true
+	}
+
+	return false
+}